ensure redcap.math.median returns float

+++ src/rios/conversion/redcap/math.py
@@ -39,7 +39,7 @@ def median(*data):
         sorted_data = sorted(data)
         n = len(sorted_data)
         if n % 2 == 1:
-            return sorted_data[n / 2]
+            return float(sorted_data[n / 2])
         else:
             m = n / 2
             return (sorted_data[m - 1] + sorted_data[m]) / 2.0
